requirements_torchcfm.txt
# DOME 主要 Python 依赖。先安装本文件，再用 tools/install_pip_env.sh
# 以 --no-deps 安装 torchcfm，避免 torchcfm 的依赖解析升级 torch/pandas。

absl-py==2.0.0
accelerate==0.30.1
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
ansi2html==1.8.0
anyio==4.0.0
apptools==5.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-lru==2.0.4
av==12.0.0
babel==2.12.1
backcall==0.2.0
black==23.11.0
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.1
comm==0.1.4
configargparse==1.7
configobj==5.0.8
contourpy==1.1.0
crcmod==1.7
cycler==0.11.0
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.7.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
diffusers==0.24.0
einops==0.7.0
entrypoints==0.4
envisage==7.0.3
executing==1.2.0
fastjsonschema==2.18.0
fire==0.5.0
flake8==5.0.4
flask==3.0.0
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.9.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.2
huggingface-hub==0.23.1
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-resources==6.0.1
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-packaging==0.12.3
jupyter-server==2.7.3
jupyter-server-fileid==0.9.1
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.7
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
lazy-loader==0.3
llvmlite==0.40.1
lyft-dataset-sdk==0.0.8
markdown==3.4.4
matplotlib==3.5.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
mmcv==2.0.1
mmdet==3.3.0
mmdet3d==1.4.0
mmengine==0.8.4
model-index==0.1.11
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.7.0
nest-asyncio==1.5.7
notebook==6.5.6
notebook-shim==0.2.3
numba==0.57.1
nuscenes-devkit==1.1.10
oauthlib==3.2.2
open3d==0.13.0
open3d-python==0.3.0.0
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.24
oss2==2.17.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.18.0
pluggy==1.3.0
plyfile==1.0.1
poethepoet==0.10.0
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.0
pot==0.9.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.9.1
pycryptodome==3.18.0
pyface==8.0.0
pyflakes==2.5.0
pyparsing==3.0.9
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pyquaternion==0.9.9
pytest==7.4.3
python-json-logger==2.0.7
pytorch-fast-transformers==0.4.0
pytorch-fid==0.3.0
pyvirtualdisplay==3.0
pywavelets==1.4.1
pyzmq==24.0.1
qtconsole==5.4.4
qtpy==2.4.0
regex==2024.5.15
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.10.2
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
send2trash==1.8.2
shapely==1.8.5
stack-data==0.6.2
swin-window-process==0.0.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
thop==0.1.1-2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.19.1
torchcde==0.2.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchmetrics==1.2.0
torchsde==0.2.6
tornado==6.3.3
tqdm==4.65.2
traitlets==5.9.0
traits==6.4.2
traitsui==8.0.0
transformers==4.41.1
trampoline==0.1.2
trimesh==4.0.4
tzdata==2023.3
uri-template==1.3.0
vtk==9.2.6
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.6.3
werkzeug==3.0.1
widgetsnbextension==4.0.8
xformers==0.0.22
y-py==0.6.2
yapf==0.40.1
ypy-websocket==0.8.4
zipp==3.16.2
